+++ NuGet/ChustaSoft.Auth.ApiKey/ITokenProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace ChustaSoft.Auth.ApiKey;
+
+public interface ITokenProvider 
+{
+    bool Validate(string token);
+}