+++ NuGet/ChustaSoft.Auth.ApiKey/changelog.md
@@ -9,4 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - XXXX-XX-XX
 ### Added
-- Added AuthenticationHandler mechanism for ApiKey
+- Added AuthenticationHandler mechanism for ApiKey Auth via lambda function as provider
+
+## [1.1.0] - XXXX-XX-XX
+### Added
+- Added generic ConfigureApiKeyAuthentication method with ITokenProvider support